+++ src/@lekoarts/gatsby-theme-emma/gatsby-plugin-theme-ui/index.js
@@ -11,6 +11,9 @@ import {
 	styles,
 	layout,
 	buttons,
+
+	cards,
+
 } from './styles/index';
 
 export default merge(tailwind, {
@@ -19,6 +22,7 @@ export default merge(tailwind, {
 	fontWeights,
 	lineHeights,
 	colors,
+	cards,
 	styles,
 	layout,
 	buttons,