+++ src/@lekoarts/gatsby-theme-emma/gatsby-plugin-theme-ui/styles/modes.js
@@ -1,9 +1,16 @@
+
+// export const modes = {
+// 	dark: {
+// 		text: 'white',
+// 	},
+// };
+
 import { tailwind } from '@theme-ui/presets';
 
 export const modes = {
 	dark: {
 		text: tailwind.colors.white,
-		//primary: tailwind.colors.pink[4],
+		primary: tailwind.colors.pink[4],
 		background: tailwind.colors.gray[8],
 		textMuted: tailwind.colors.gray[5],
 	},