+++ src/@lekoarts/gatsby-theme-emma/gatsby-plugin-theme-ui/styles/styles.js
@@ -1,62 +1,21 @@
-export const styles1 = {
-	root: {
-		margin: 0,
-		padding: 0,
-		color: 'text',
-		boxSizing: 'border-box',
-		backgroundColor: 'background',
-		MozOsxFontSmoothing: 'grayscale',
-		WebkitFontSmoothing: 'antialiased',
-		textRendering: 'optimizeLegibility',
-	},
-	p: {
-		fontSize: 9,
-		letterSpacing: '-0.003em',
-		lineHeight: 'body',
-		'--baseline-multiplier': 0.179,
-		'--x-height-multiplier': 0.35,
-	},
-	h1: {
-		fontSize: 5,
-		mt: 2,
-	},
-	h2: {
-		fontSize: 6,
-		mt: 2,
-	},
-	h3: {
-		fontSize: 7,
-		mt: 3,
-	},
-	h4: {
-		fontSize: 8,
-	},
-	h5: {
-		fontSize: 9,
-	},
-	h6: {
-		fontSize: 10,
-		mb: 2,
-	},
-};
 
 export const styles = {
 	root: {
-		color: 'text',
-		backgroundColor: 'background',
 		margin: 0,
 		padding: 0,
+		color: 'gs800',
+		backgroundColor: 'bg',
 		boxSizing: 'border-box',
-		textRendering: 'optimizeLegibility',
-		WebkitFontSmoothing: 'antialiased',
 		MozOsxFontSmoothing: 'grayscale',
+		WebkitFontSmoothing: 'antialiased',
+		textRendering: 'optimizeLegibility',
 	},
 	p: {
 		fontSize: [1, 2],
-		letterSpacing: '-0.003em',
 		lineHeight: 'body',
-		'--baseline-multiplier': 0.179,
+		letterSpacing: '-0.003em',
 		'--x-height-multiplier': 0.35,
+		'--baseline-multiplier': 0.179,
 	},
 	h1: {
 		fontSize: [5, 6],